refactor: change the name of the import
